Simplify styles, inline button and move panel file
