--- ui/src/app/version-info/components/version-info.scss
@@ -1,23 +1,0 @@
-.version-info-btn {
-    white-space: nowrap;
-    cursor: pointer;
-
-    &:hover, &:focus {
-        text-decoration: underline;
-    }
-}
-
-.version-copy-btn-container {
-    margin-top: 15px;
-    display: flex;
-    flex-direction: row-reverse;
-
-    & button {
-        min-width: 20ch;
-        & i {
-            text-align: center;
-            margin-right: 0.5em;
-            min-width: 1rem;
-        }
-    }
-}
